# swagger.json
{
    "swagger": "2.0",
    "host": "localhost:3000",
    "basePath": "/api",
    "info": {
      "description": "Rest API para las plataformas",
      "version": "1.0.0",
      "title": "API Pruebas",
      "contact": {
        "email": "emiliocaraballo9810@gmail.com"
      }
    },
...
    "schemes": [
      "http",
      "https"
    ],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
      "/service": {
        "get": {
          "tags": [
            "services"
          ],
          "summary": "Obtener listado de servicios",
          "operationId": "getservice",
          "description": "En esta listado cada item tiene su id y titulo el id es importante al momento de realizar la implementacion\n de la api/order ya que en esta parte se necesita agregar la columna service y aqui se agrega el id de servicios.\n",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "la solicitud se ha realizado correctamente.",
              "schema": {
                "$ref": "#/definitions/Service"
              }
            },
            "400": {
              "description": "El servidor no pudo entender la solicitud debido a una sintaxis incorrecta.",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            },
            "404": {
              "description": "Recurso no fue encontrado.",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            }          }
        },
        "post": {
          "tags": [
            "services"
          ],
          "summary": "Crear Servicios",
          "operationId": "createService",
          "description": "Se crea un servicio nuevo y se retorna un json indicado si se registro con exito o no\n",
          "produces": [
            "application/json"
          ],
          "security": [
            {
              "Bearer": []
            }
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "schema": {
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "titulo"
                  },
                  "description": {
                    "type": "string",
                    "example": "Descripción"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "se ha realizado correctamente y, como resultado, se ha creado un nuevo recurso.",
              "schema": {
                "$ref": "#/definitions/Success"
              }
            },
            "400": {
              "description": "El servidor no pudo entender la solicitud debido a una sintaxis incorrecta.",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
      "Service": {
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "code": {
            "type": "number",
            "example":"1"
          },  
          "service":{
            "type":"array",
            "items":{
              "type":"object",
              "example":
              {
                "id":1,
                "title":"Service #1",
                "description":"Pruebas",
                "create_at":"2022-01-25 11:51:56",
                "update_at":"2022-01-25 11:51:56"
              }
            }
          }
        }
      },
...
      "Success":{
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "number",
            "example": "1"
          },
          "message": {
            "type": "string",
            "example": "Se ha registrado exitosamente"
          }
        }
      },
...
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "properties": {
          "code": {
            "type": "number",
            "example": "0"
          },
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      }
    }
  }